+++ apps/member-profile/app/routes/onboarding.emails_.verify.tsx
@@ -5,15 +5,8 @@ import {
   redirect,
 } from '@remix-run/node';
 import { Form, useActionData, useLoaderData } from '@remix-run/react';
-import { type z } from 'zod';
 
-import { job } from '@oyster/core/bull';
-import {
-  OneTimeCode,
-  OneTimeCodePurpose,
-} from '@oyster/core/member-profile/ui';
-import { db } from '@oyster/db';
-import { StudentEmail } from '@oyster/types';
+import { addEmail, AddEmailInput } from '@oyster/core/member-profile/server';
 import {
   ErrorMessage,
   Field,
@@ -26,7 +19,6 @@ import {
   OnboardingBackButton,
   OnboardingButtonGroup,
   OnboardingContinueButton,
-  OnboardingSectionDescription,
   OnboardingSectionTitle,
 } from '@/routes/onboarding';
 import { Route } from '@/shared/constants';
@@ -36,40 +28,28 @@ import { ensureUserAuthenticated, user } from '@/shared/session.server';
 export async function loader({ request }: LoaderFunctionArgs) {
   await ensureUserAuthenticated(request);
 
-  const email = await addEmailCookie.parse(request.headers.get('Cookie'));
+  const email = await getEmailFromCookie(request);
 
   if (!email) {
-    return redirect(Route['/profile/emails/add/start']);
+    return redirect(Route['/onboarding/emails']);
   }
 
   return json({ email });
 }
 
-const AddEmailInput = StudentEmail.pick({
-  email: true,
-  studentId: true,
-})
-  .extend({ code: OneTimeCode.shape.value })
-  .required();
-
-type AddEmailInput = z.infer<typeof AddEmailInput>;
-
-const AddEmailFormData = AddEmailInput.pick({
-  code: true,
-});
-
-type AddEmailFormData = z.infer<typeof AddEmailFormData>;
-
 export async function action({ request }: ActionFunctionArgs) {
   const session = await ensureUserAuthenticated(request);
 
-  const { data, errors, ok } = await validateForm(request, AddEmailFormData);
+  const { data, errors, ok } = await validateForm(
+    request,
+    AddEmailInput.pick({ code: true })
+  );
 
   if (!ok) {
     return json({ errors }, { status: 400 });
   }
 
-  const email = await addEmailCookie.parse(request.headers.get('Cookie'));
+  const email = await getEmailFromCookie(request);
 
   if (!email) {
     return json({
@@ -91,56 +71,11 @@ export async function action({ request }: ActionFunctionArgs) {
   }
 }
 
-async function addEmail(input: AddEmailInput) {
-  const existingEmail = await db
-    .selectFrom('studentEmails')
-    .select(['studentId'])
-    .where('email', 'ilike', input.email)
-    .executeTakeFirst();
-
-  if (existingEmail) {
-    throw new Error(
-      existingEmail.studentId === input.studentId
-        ? 'This email already belongs to you.'
-        : 'The email you are trying to add belongs to another member.'
-    );
-  }
-
-  const oneTimeCode = await db
-    .selectFrom('oneTimeCodes')
-    .select('id')
-    .where('email', 'ilike', input.email)
-    .where('purpose', '=', OneTimeCodePurpose.ADD_STUDENT_EMAIL)
-    .where('studentId', '=', input.studentId)
-    .where('value', '=', input.code as string)
-    .executeTakeFirst();
-
-  if (!oneTimeCode) {
-    throw new Error('The code was either wrong or expired. Please try again.');
-  }
-
-  await db.transaction().execute(async (trx) => {
-    await trx
-      .insertInto('studentEmails')
-      .values({
-        email: input.email,
-        studentId: input.studentId,
-      })
-      .execute();
-
-    await trx
-      .deleteFrom('oneTimeCodes')
-      .where('id', '=', oneTimeCode.id)
-      .execute();
-  });
-
-  job('member_email.added', {
-    email: input.email,
-    studentId: input.studentId,
-  });
+async function getEmailFromCookie(request: Request) {
+  return addEmailCookie.parse(request.headers.get('Cookie'));
 }
 
-export default function OnboardingEmailForm() {
+export default function OnboardingVerifyEmailForm() {
   const { email } = useLoaderData<typeof loader>();
   const actionData = useActionData<typeof action>();
   const { error, errors } = getErrors(actionData);
@@ -148,13 +83,14 @@ export default function OnboardingEmailForm() {
   return (
     <Form className="form" method="post">
       <OnboardingSectionTitle>Email Addresses</OnboardingSectionTitle>
-      <OnboardingSectionDescription>
-        Please input the 6-digit passcode that you received to complete the
-        addition of <span style={{ fontWeight: 700 }}>{email}</span> to your
-        profile.
-      </OnboardingSectionDescription>
 
-      <Field error={errors.code} labelFor="code" required>
+      <Field
+        description={`Please input the 6-digit passcode sent to ${email}.`}
+        error={errors.code}
+        label="One-Time Code"
+        labelFor="code"
+        required
+      >
         <Input autoFocus id="code" name="code" required />
       </Field>
